FY-43/refactor: 토큰 fetch 로직 AuthProvider 내부로 옮김
